cc: refactor DeepTensor for multiple-backend framework
See #3119
